added toString for all model classes, and moved one of the stub classes into stub folder bc it was not in it aleady
