+++ src/Stubs/PreviousTimesheetsStub.java
@@ -1,5 +1,7 @@
-package model;
-import java.util.*; 
+package Stubs;
+import java.util.*;
+
+import model.PreviousTimesheets; 
 
 public class PreviousTimesheetsStub extends PreviousTimesheets {
